Gave Interpolator base class more responsibility
The evaluation of the interpolation is handled more  by the base class,
to make sure bounds checking is performed. The concrete classes also get
the segment of the interpolation passed to the now, rather than needing
to find it themselves. Additionally, the interpolation is evaluated with
the 'eval' method, rather than with the () operator.